# common-docs/openapi/openapi.v1.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Aini-Inu API",
        "description": "Aini-Inu \ubc31\uc5d4\ub4dc REST API \ubb38\uc11c",
        "contact": {
            "name": "Aini-Inu Backend Team"
        },
        "version": "v1"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Local"
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ],
...
    "paths": {
...
        "/api/v1/auth/refresh": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "\ud1a0\ud070 \uac31\uc2e0",
                "description": "Refresh Token\uc744 \uac80\uc99d\ud574 \uc0c8\ub85c\uc6b4 Access/Refresh Token\uc744 \ubc1c\uae09\ud569\ub2c8\ub2e4.",
                "operationId": "refresh",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TokenRefreshRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "\uac31\uc2e0 \uc131\uacf5",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseLoginResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "\uc694\uccad \uac80\uc99d \uc2e4\ud328",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseLoginResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "\uc720\ud6a8\ud558\uc9c0 \uc54a\uac70\ub098 \ub9cc\ub8cc\ub41c \ud1a0\ud070",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseLoginResponse"
                                }
                            }
                        }
                    }
                },
                "security": []
            }
        },
        "/api/v1/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "\ub85c\uadf8\uc544\uc6c3",
                "description": "Refresh Token\uc744 \ud3d0\uae30\ud574 \uc7ac\ubc1c\uae09\uc744 \ucc28\ub2e8\ud569\ub2c8\ub2e4.",
                "operationId": "logout",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TokenRevokeRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "\ub85c\uadf8\uc544\uc6c3 \uc131\uacf5",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseVoid"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "\uc694\uccad \uac80\uc99d \uc2e4\ud328",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseVoid"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "\uc778\uc99d \uc2e4\ud328",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseVoid"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "\uc774\uba54\uc77c \ub85c\uadf8\uc778",
                "description": "\uc774\uba54\uc77c/\ube44\ubc00\ubc88\ud638\ub85c \ub85c\uadf8\uc778\ud558\uace0 Access/Refresh Token\uc744 \ubc1c\uae09\ud569\ub2c8\ub2e4.",
                "operationId": "login",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AuthLoginRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "\ub85c\uadf8\uc778 \uc131\uacf5",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseLoginResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "\uc694\uccad \uac80\uc99d \uc2e4\ud328",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseLoginResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "\uc778\uc99d \uc2e4\ud328",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseLoginResponse"
                                }
                            }
                        }
                    }
                },
                "security": []
            }
        },
...
    },
    "components": {
        "schemas": {
            "ApiResponseVoid": {
                "type": "object",
                "description": "\uacf5\ud1b5 API \uc751\ub2f5 \ub798\ud37c",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "description": "\uc694\uccad \uc131\uacf5 \uc5ec\ubd80",
                        "example": true
                    },
                    "status": {
                        "type": "integer",
                        "format": "int32",
                        "description": "HTTP \uc0c1\ud0dc \ucf54\ub4dc",
                        "example": 200
                    },
                    "data": {
                        "type": "object",
                        "description": "\uc2e4\uc81c \uc751\ub2f5 \ub370\uc774\ud130"
                    },
                    "errorCode": {
                        "type": "string",
                        "description": "\uc5d0\ub7ec \ucf54\ub4dc (\uc131\uacf5 \uc2dc null)",
                        "example": "C101"
                    },
                    "message": {
                        "type": "string",
                        "description": "\uc5d0\ub7ec \uba54\uc2dc\uc9c0 (\uc131\uacf5 \uc2dc null)",
                        "example": "\uc778\uc99d\uc774 \ud544\uc694\ud569\ub2c8\ub2e4"
                    }
                }
            },
...
            "ApiResponseLoginResponse": {
                "type": "object",
                "description": "\uacf5\ud1b5 API \uc751\ub2f5 \ub798\ud37c",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "description": "\uc694\uccad \uc131\uacf5 \uc5ec\ubd80",
                        "example": true
                    },
                    "status": {
                        "type": "integer",
                        "format": "int32",
                        "description": "HTTP \uc0c1\ud0dc \ucf54\ub4dc",
                        "example": 200
                    },
                    "data": {
                        "$ref": "#/components/schemas/LoginResponse",
                        "description": "\uc2e4\uc81c \uc751\ub2f5 \ub370\uc774\ud130"
                    },
                    "errorCode": {
                        "type": "string",
                        "description": "\uc5d0\ub7ec \ucf54\ub4dc (\uc131\uacf5 \uc2dc null)",
                        "example": "C101"
                    },
                    "message": {
                        "type": "string",
                        "description": "\uc5d0\ub7ec \uba54\uc2dc\uc9c0 (\uc131\uacf5 \uc2dc null)",
                        "example": "\uc778\uc99d\uc774 \ud544\uc694\ud569\ub2c8\ub2e4"
                    }
                }
            },
            "LoginResponse": {
                "type": "object",
                "properties": {
                    "accessToken": {
                        "type": "string"
                    },
                    "refreshToken": {
                        "type": "string"
                    },
                    "tokenType": {
                        "type": "string"
                    },
                    "expiresIn": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "memberId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "newMember": {
                        "type": "boolean"
                    },
                    "isNewMember": {
                        "type": "boolean"
                    }
                }
            },
...
            "TokenRefreshRequest": {
                "type": "object",
                "properties": {
                    "refreshToken": {
                        "type": "string",
                        "minLength": 1
                    }
                },
                "required": [
                    "refreshToken"
                ]
            },
            "TokenRevokeRequest": {
                "type": "object",
                "properties": {
                    "refreshToken": {
                        "type": "string",
                        "minLength": 1
                    }
                },
                "required": [
                    "refreshToken"
                ]
            },
            "AuthLoginRequest": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string",
                        "minLength": 1
                    },
                    "password": {
                        "type": "string",
                        "minLength": 1
                    }
                },
                "required": [
                    "email",
                    "password"
                ]
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}